Use npx ts-node as a workaround to compile (HAS BUGS! See details)
Command npm run build exits with errors presumably related to dependencies and typings.
We have attempted to move tsconfig.common.ts from the project's root to the package. This makes npm run build execute with success; however, we cannot use this solution as it breaks modularity.
In other words, it seems that the culprit of the problem is
"extends": "../../tsconfig.common.json"

As a workaround, Carlo suggested we temporarily use:
npx ts-node -r tsconfig-paths/register src/bootstrap.ts

Signed-off-by: Andrea Stedile <andrea.stedile@studenti.unitn.it>